+++ src/rules/comment-whitespace-inside/__tests__/index.js
@@ -18,6 +18,8 @@ testRule("always", tr => {
   tr.ok("/**** comment ***/")
   tr.ok("/*\ncomment\n*/")
   tr.ok("/*\tcomment   */")
+  tr.ok("/*! copyright */")
+  tr.ok("/*# sourcemap */")
 
   tr.notOk("/*comment */", {
     message: messages.expectedOpening,
@@ -54,6 +56,11 @@ testRule("always", tr => {
     line: 3,
     column: 7,
   })
+  tr.notOk("/*!copyright */", {
+    message: messages.expectedOpening,
+    line: 1,
+    column: 3,
+  })
 })
 
 testRule("never", tr => {
@@ -65,6 +72,8 @@ testRule("never", tr => {
   tr.ok("/*comment\n\ncomment*/")
   tr.ok("/**comment*/")
   tr.ok("/****comment***/")
+  tr.ok("/*! copyright */")
+  tr.ok("/*# sourcemap */")
 
   tr.notOk("/* comment*/", {
     message: messages.rejectedOpening,